+++ NEWS.md
@@ -4,6 +4,7 @@
 -   Now, **ody_summarise_df** plots the density curves following the same order as shown in the table and if the test fails when comparing groups, the empty results table is not shown.
 -   Minor modification of the *quality* folder structure. Now, it only contains the folders *verification* and *tables.*
 -   Added the argument *export* to **ody_add_to_datasets**. If *export* = TRUE, the dataset is exported as csv to quality/tables.
+-   Added **Update odytools** to add-ins to install the last master or development version of odytools.
 
 # odytools 0.7.0